refactor: move business logic from controllers to dedicated service layer
