+++ src/main/java/nl/pkock/brewhub_backend/config/WebConfig.java
@@ -1,4 +1,4 @@
-package nl.pkock.brewhub_backend;
+package nl.pkock.brewhub_backend.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
