+++ src/main/java/nl/pkock/brewhub_backend/rating/controller/RatingController.java
@@ -1,53 +1,24 @@
 package nl.pkock.brewhub_backend.rating.controller;
 
 import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
 import nl.pkock.brewhub_backend.rating.dto.CreateRatingRequest;
 import nl.pkock.brewhub_backend.rating.dto.RatingDTO;
-import nl.pkock.brewhub_backend.order.models.Order;
-import nl.pkock.brewhub_backend.order.models.OrderStatus;
-import nl.pkock.brewhub_backend.rating.model.Rating;
-import nl.pkock.brewhub_backend.auth.models.User;
-import nl.pkock.brewhub_backend.order.repository.OrderRepository;
-import nl.pkock.brewhub_backend.rating.repository.RatingRepository;
-import nl.pkock.brewhub_backend.auth.repository.UserRepository;
 import nl.pkock.brewhub_backend.auth.security.UserPrincipal;
+import nl.pkock.brewhub_backend.rating.service.RatingService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.math.BigDecimal;
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api")
+@RequiredArgsConstructor
 public class RatingController {
-
-    private final RatingRepository ratingRepository;
-    private final UserRepository userRepository;
-    private final OrderRepository orderRepository;
-
-    public RatingController(RatingRepository ratingRepository, UserRepository userRepository, OrderRepository orderRepository) {
-        this.ratingRepository = ratingRepository;
-        this.userRepository = userRepository;
-        this.orderRepository = orderRepository;
-    }
-
-    private RatingDTO convertToDTO(Rating rating) {
-        RatingDTO dto = new RatingDTO();
-        dto.setId(rating.getId());
-        dto.setCustomerName(rating.getCustomer().getFirstName() + " " + rating.getCustomer().getLastName());
-        dto.setScore(rating.getScore());
-        dto.setComment(rating.getComment());
-        dto.setCreatedAt(rating.getCreatedAt());
-        if (rating.getOrder() != null) {
-            dto.setOrderId(rating.getOrder().getId());
-        }
-        return dto;
-    }
+    private final RatingService ratingService;
 
     @PostMapping("/user/ratings/{retailerId}/order/{orderId}")
     @PreAuthorize("hasRole('USER')")
@@ -56,80 +27,24 @@ public class RatingController {
             @PathVariable Long retailerId,
             @PathVariable Long orderId,
             @Valid @RequestBody CreateRatingRequest request) {
-
         UserPrincipal userPrincipal = (UserPrincipal) authentication.getPrincipal();
-        User customer = userRepository.findById(userPrincipal.getId())
-                .orElseThrow(() -> new RuntimeException("User not found"));
-
-        User retailer = userRepository.findById(retailerId)
-                .orElseThrow(() -> new RuntimeException("Retailer not found"));
-
-        Order order = orderRepository.findById(orderId)
-                .orElseThrow(() -> new RuntimeException("Order not found"));
-
-        // Validate order belongs to user and retailer
-        if (!order.getCustomer().getId().equals(customer.getId()) ||
-                !order.getRetailer().getId().equals(retailerId)) {
-            throw new RuntimeException("Invalid order for this rating");
-        }
-
-        // Validate order status
-        if (order.getStatus() != OrderStatus.DELIVERED) {
-            throw new RuntimeException("Can only rate delivered orders");
-        }
-
-        // Check if order already has a rating
-        if (ratingRepository.findByOrderId(orderId).isPresent()) {
-            throw new RuntimeException("Order already has a rating");
-        }
-
-        Rating rating = new Rating();
-        rating.setRetailer(retailer);
-        rating.setCustomer(customer);
-        rating.setOrder(order);
-        rating.setScore(request.getScore());
-        rating.setComment(request.getComment());
-        rating.setCreatedAt(LocalDateTime.now());
-
-        Rating savedRating = ratingRepository.save(rating);
-
-        // Update retailer's average rating
-        Double avgRating = ratingRepository.findAverageRatingByRetailerId(retailerId)
-                .orElse(0.0);
-
-        retailer.setAverageRating(BigDecimal.valueOf(avgRating));
-        retailer.setTotalRatings(retailer.getTotalRatings() == null ? 1 : retailer.getTotalRatings() + 1);
-        userRepository.save(retailer);
-
-        return ResponseEntity.ok(convertToDTO(savedRating));
+        RatingDTO rating = ratingService.createRating(userPrincipal.getId(), retailerId, orderId, request);
+        return ResponseEntity.ok(rating);
     }
 
     @GetMapping("/public/retailers/{retailerId}/ratings")
-    @Transactional(readOnly = true)
     public ResponseEntity<List<RatingDTO>> getRetailerRatings(@PathVariable Long retailerId) {
-        List<Rating> ratings = ratingRepository.findByRetailerId(retailerId);
-        List<RatingDTO> ratingDTOs = ratings.stream()
-                .map(this::convertToDTO)
-                .collect(Collectors.toList());
-        return ResponseEntity.ok(ratingDTOs);
+        return ResponseEntity.ok(ratingService.getRetailerRatings(retailerId));
     }
 
     @GetMapping("/public/retailers/{retailerId}/average-rating")
     public ResponseEntity<BigDecimal> getRetailerAverageRating(@PathVariable Long retailerId) {
-        User retailer = userRepository.findById(retailerId)
-                .orElseThrow(() -> new RuntimeException("Retailer not found"));
-        return ResponseEntity.ok(retailer.getAverageRating());
+        return ResponseEntity.ok(ratingService.getRetailerAverageRating(retailerId));
     }
 
     @GetMapping("/public/orders/{orderId}/rating")
     public ResponseEntity<RatingDTO> getOrderRating(@PathVariable Long orderId) {
-        Rating rating = ratingRepository.findByOrderId(orderId)
-                .orElse(null);
-
-        if (rating == null) {
-            return ResponseEntity.notFound().build();
-        }
-
-        return ResponseEntity.ok(convertToDTO(rating));
+        RatingDTO rating = ratingService.getOrderRating(orderId);
+        return rating != null ? ResponseEntity.ok(rating) : ResponseEntity.notFound().build();
     }
 }
